# api/openapi/openapi.gen.yaml
...
openapi: 3.1.0
info:
  title: Illarin API
  version: 0.1.0
  description: >-
    State-changing requests that use the session cookie also carry Illarin's
    browser-mutation proof and exact browser origin.
servers:
  - url: /api
paths:
  /v1/account/discord:
    delete:
      operationId: detachDiscord
      security:
        - sessionCookie: []
          browserMutation: []
      responses:
        '200':
          description: The account with Discord detached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: No account is signed in
        '409':
          description: Discord is absent or is the last usable sign-in method
  /v1/account/email:
    patch:
      operationId: changeUnverifiedEmail
      security:
        - sessionCookie: []
          browserMutation: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmailRequest'
      responses:
        '200':
          description: The account with a fresh verification link sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: The address is not valid
        '401':
          description: No account is signed in
        '409':
          description: The address is verified here or elsewhere
  /v1/account/handle:
    patch:
      operationId: renameHandle
      security:
        - sessionCookie: []
          browserMutation: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameHandleRequest'
      responses:
        '200':
          description: The account under its new handle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: The handle is not valid
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
        '409':
          description: The handle is active or retired
  /v1/account/password:
    put:
      operationId: setPassword
      security:
        - sessionCookie: []
          browserMutation: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRequest'
      responses:
        '200':
          description: The account with a password set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: The password is empty
        '401':
          description: No account is signed in
        '409':
          description: The account already has a password
  /v1/account/nsfw-visibility:
    put:
      operationId: setNsfwVisibility
      security:
        - sessionCookie: []
          browserMutation: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NsfwVisibilityRequest'
      responses:
        '204':
          description: The reader's content preference is saved
        '400':
          description: The visibility value is invalid
        '401':
          description: No account is signed in
  /v1/account/profile:
    put:
      operationId: savePublicProfile
      security:
        - sessionCookie: []
          browserMutation: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveProfileRequest'
      responses:
        '200':
          description: The public profile as visitors now see it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: A profile field is not valid
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
  /v1/account/profile/avatar:
    put:
      operationId: setProfileAvatar
      security:
        - sessionCookie: []
          browserMutation: []
      requestBody:
        required: true
        description: Send the image as the only form data part.
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: The public profile carrying its new avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: The image is not valid
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
        '413':
          description: The image exceeds the upload limit
        '503':
          description: The storage reserve cannot accept the image
    delete:
      operationId: removeProfileAvatar
      security:
        - sessionCookie: []
          browserMutation: []
      responses:
        '200':
          description: The public profile without an avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
        '404':
          description: The profile has no avatar
  /v1/auth/sign-up:
    post:
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: The new signed-in account
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: The account details are not valid
        '409':
          description: The handle or verified address is already claimed
        '429':
          description: Too many sign-up attempts came from this source
  /v1/auth/sign-in:
    post:
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: The signed-in account
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: The credentials do not identify one account
        '429':
          description: Too many sign-in attempts came from this source
  /v1/auth/discord:
    get:
      operationId: beginDiscord
      parameters:
        - name: intent
          in: query
          schema:
            type: string
            enum:
              - sign-in
              - attach
            default: sign-in
        - name: returnTo
          in: query
          description: Internal page to resume after sign-in
          schema:
            type: string
      responses:
        '303':
          description: Continue on Discord
        '401':
          description: Attaching Discord requires a signed-in account
        '403':
          description: Attaching Discord requires a verified account
        '503':
          description: Discord sign-in is not configured
  /v1/auth/discord/callback:
    get:
      operationId: completeDiscord
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        '303':
          description: Continue on Illarin
...
  /v1/auth/verify-email:
    post:
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: The account with its address verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: The link is invalid
          expired or has already lost the address: null
        '409':
          description: Another account verified the address first
  /v1/auth/password-reset:
    post:
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '204':
          description: A reset is sent when the verified account exists
        '400':
          description: The request is not valid JSON
        '429':
          description: Too many reset requests came from this source
  /v1/auth/password-reset/complete:
    post:
      operationId: completePasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletePasswordResetRequest'
      responses:
        '204':
          description: The password is set
        '400':
          description: The link or password is invalid
        '429':
          description: Too many reset attempts came from this source
  /v1/link/requests:
    post:
      operationId: startLinkRequest
      description: >-
        Start Illarin's registration-free device fallback for an installation
        that cannot receive a loopback callback. It follows RFC 8628 security
        and polling semantics but is not an OAuth token endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartLinkRequest'
      responses:
        '201':
          description: The code the installation keeps and the code it shows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkRequest'
        '400':
          description: The declaration or requested scopes are not valid
        '413':
          description: The request body is too large
        '429':
          description: Too many link requests came from this source
  /v1/link/poll:
    post:
      operationId: pollLinkRequest
      description: >-
        Ask whether the creator has approved yet. Wait at least the interval the
        link request came with between calls.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollLinkRequest'
      responses:
        '200':
          description: Still waiting, or linked with a new token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkPollResult'
        '400':
          description: The request was denied or expired
        '404':
          description: The request was redeemed or never existed
        '413':
          description: The request body is too large
        '429':
          description: Slow down before polling again
          headers:
            Retry-After:
              schema:
                type: integer
  /v1/link/requests/{userCode}:
    get:
      operationId: getLinkRequest
      description: What the installation behind a user code is asking for.
      security:
        - sessionCookie: []
      parameters:
        - name: userCode
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserCode'
      responses:
        '200':
          description: The unverified names, declaration and scopes to review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingDeviceLink'
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
        '404':
          description: No unapproved link request has that code
        '429':
          description: Too many codes were entered
  /v1/link/requests/{userCode}/approve:
    post:
      operationId: approveLinkRequest
      description: >-
        Approve a link request. A link request may be approved once and redeemed
        once. This browser-only operation requires Illarin's request header.
      security:
        - sessionCookie: []
          browserMutation: []
      parameters:
        - name: userCode
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserCode'
        - $ref: '#/components/parameters/IllarinRequest'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceLinkDecision'
      responses:
        '200':
          description: What was approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingLink'
        '400':
          description: A path or browser-proof header parameter is malformed
        '401':
          description: No account is signed in
        '403':
          description: Email is unverified or the browser origin/request proof is invalid
        '404':
          description: No unapproved link request has that code
  /v1/link/requests/{userCode}/deny:
    post:
      operationId: denyLinkRequest
      description: Deny the reviewed device request and stop its polling installation.
      security:
        - sessionCookie: []
          browserMutation: []
      parameters:
        - name: userCode
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserCode'
        - $ref: '#/components/parameters/IllarinRequest'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceLinkDecision'
      responses:
        '204':
          description: The request is denied
        '400':
          description: A path or browser-proof header parameter is malformed
        '401':
          description: No account is signed in
        '403':
          description: Email is unverified or the browser origin/request proof is invalid
        '404':
          description: No pending reviewed request has that code
  /v1/link/authorizations:
    post:
      operationId: startLinkAuthorization
      description: >-
        Start same-device browser authorization with S256 PKCE. Only an exact
        IPv4 or IPv6 loopback callback is accepted; its port may vary.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartLinkAuthorization'
      responses:
        '201':
          description: The short-lived Illarin URL to open in the system browser
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkAuthorization'
        '400':
          description: The callback
          PKCE data or declaration is not valid: null
        '413':
          description: The request body is too large
        '429':
          description: Too many link requests came from this source
  /v1/link/authorizations/{requestCode}:
    get:
      operationId: getLinkAuthorization
      description: What the installation behind a browser authorization is asking for.
      security:
        - sessionCookie: []
      parameters:
        - name: requestCode
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/RequestCode'
      responses:
        '200':
          description: The unverified names, declaration and scopes to review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingLink'
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
        '404':
          description: No pending authorization has that request code
  /v1/link/authorizations/{requestCode}/approve:
    post:
      operationId: approveLinkAuthorization
      description: Approve once and return the validated loopback redirect.
      security:
        - sessionCookie: []
          browserMutation: []
      parameters:
        - name: requestCode
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/RequestCode'
        - $ref: '#/components/parameters/IllarinRequest'
      responses:
        '200':
          description: The exact loopback URL carrying the one-use code and state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkRedirect'
        '400':
          description: A path or browser-proof header parameter is malformed
        '401':
          description: No account is signed in
        '403':
          description: Email is unverified or the browser origin/request proof is invalid
        '404':
          description: No pending authorization has that request code
  /v1/link/authorizations/{requestCode}/deny:
    post:
      operationId: denyLinkAuthorization
      description: Deny once and return the validated loopback error redirect.
      security:
        - sessionCookie: []
          browserMutation: []
      parameters:
        - name: requestCode
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/RequestCode'
        - $ref: '#/components/parameters/IllarinRequest'
      responses:
        '200':
          description: The exact loopback URL carrying access_denied and state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkRedirect'
        '400':
          description: A path or browser-proof header parameter is malformed
        '401':
          description: No account is signed in
        '403':
          description: Email is unverified or the browser origin/request proof is invalid
        '404':
          description: No pending authorization has that request code
  /v1/link/token:
    post:
      operationId: exchangeLinkAuthorization
      description: Exchange one approved authorization code using its PKCE verifier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeLinkAuthorization'
      responses:
        '200':
          description: A short-lived access token and rotating refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceTokenGrant'
        '400':
          description: The code
          verifier or callback is not valid: null
        '413':
          description: The request body is too large
        '429':
          description: Too many exchanges came from this source
  /v1/link/refresh:
    post:
      operationId: refreshInstanceToken
      description: >-
        Rotate one live refresh token. Replaying a replacement retained in the
        90-day detection window revokes that instance and every access token it
        holds. A family also expires after 90 days without authenticated use.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshInstanceToken'
      responses:
        '200':
          description: A replacement access and refresh token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceTokenGrant'
        '401':
          description: The refresh token is not live
        '413':
          description: The request body is too large
        '429':
          description: Too many refresh attempts came from this source
  /v1/instances:
    get:
      operationId: listInstances
      description: >-
        Every instance linked to the signed-in account, live ones first, then
        the record of the ones they revoked.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The creator's linked instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedInstanceList'
        '401':
          description: No account is signed in
  /v1/instances/me:
    get:
      operationId: getInstance
      description: >-
        The instance the access token belongs to. Every authenticated request
        updates lastSeenAt, and this one exists so an installation can check its
        token without sending anything.
      security:
        - instanceToken: []
      responses:
        '200':
          description: The instance holding the credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedInstance'
        '401':
          description: The access token does not identify a live instance
    put:
      operationId: updateInstance
      description: >-
        Replace the live instance's non-authoritative interoperability
        declaration. This cannot change its granted scopes.
      security:
        - instanceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInstance'
      responses:
        '200':
          description: The updated instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedInstance'
        '400':
          description: The declaration is not valid
        '401':
          description: The access token does not identify a live instance
        '413':
          description: The request body is too large
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: illarin_session
    browserMutation:
      type: apiKey
      in: header
      name: X-Illarin-Request
      description: >-
        Required with the session cookie on state-changing browser requests. The
        value must be 1 and the Origin must be Illarin's browser origin.
    instanceToken:
      type: http
      scheme: bearer
      description: A short-lived access token issued to one linked instance.
...
  schemas:
...
    Account:
      type: object
      required:
        - id
        - handle
        - email
        - emailVerified
        - discordLinked
        - hasPassword
        - role
      properties:
        id:
          type: string
          format: uuid
        handle:
          type: string
        email:
          type:
            - string
            - 'null'
          format: email
        emailVerified:
          type: boolean
        discordLinked:
          type: boolean
        hasPassword:
          type: boolean
        role:
          type: string
          enum:
            - user
            - moderator
            - admin
    ChangeEmailRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
...
    RenameHandleRequest:
      type: object
      required:
        - handle
      properties:
        handle:
          type: string
          minLength: 3
          maxLength: 32
...
    PasswordRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          minLength: 1
...